requirements.txt
numpy
scipy
pandas
matplotlib
mne
tensorflow-macos
scikit-learn


